1.20.6

+++ common/src/main/java/com/matthewperiut/clay/network/client/ClientNetworkHandler.java
@@ -2,48 +2,47 @@ package com.matthewperiut.clay.network.client;
 
 import com.matthewperiut.clay.ClayMod;
 import com.matthewperiut.clay.entity.soldier.SoldierDollEntity;
-import com.matthewperiut.clay.registry.UpgradeRegistry;
+import com.matthewperiut.clay.network.payload.SyncUpgradesPayload;
 import com.matthewperiut.clay.upgrade.ISoldierUpgrade;
 import dev.architectury.networking.NetworkManager;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.entity.Entity;
-import net.minecraft.network.PacketByteBuf;
-import net.minecraft.util.Identifier;
 
+import java.util.List;
 import java.util.Optional;
-
-import static com.matthewperiut.clay.network.UpgradeAdded.UPGRADE_ADDED_PACKET;
-import static com.matthewperiut.clay.network.UpgradeRemoved.UPGRADE_REMOVED_PACKET;
 
 @Environment(EnvType.CLIENT)
 public class ClientNetworkHandler {
     public static void init() {
-        NetworkManager.registerReceiver(NetworkManager.Side.S2C, UPGRADE_ADDED_PACKET, (buf, context) -> {
-            Optional<SoldierUpgradeContainer> optionalSoldier = handleIncomingUpgradePacket(buf, context);
-            if (optionalSoldier.isEmpty())
-                return;
-            SoldierUpgradeContainer container = optionalSoldier.get();
-            container.soldier.upgrades.add(container.upgrade);
-            container.upgrade.onAdd(container.soldier);
-        });
-        NetworkManager.registerReceiver(NetworkManager.Side.S2C, UPGRADE_REMOVED_PACKET, (buf, context) -> {
-            Optional<SoldierUpgradeContainer> optionalSoldier = handleIncomingUpgradePacket(buf, context);
-            if (optionalSoldier.isEmpty())
-                return;
-            SoldierUpgradeContainer container = optionalSoldier.get();
-            container.soldier.upgrades.remove(container.upgrade);
-            container.upgrade.onRemove(container.soldier);
-
+        NetworkManager.registerReceiver(NetworkManager.Side.S2C, SyncUpgradesPayload.ID, SyncUpgradesPayload.CODEC, (buf, context) -> {
+            if (buf.action() == SyncUpgradesPayload.Action.ADD) {
+                List<ISoldierUpgrade> upgrades = buf.upgrades();
+                for (int i = 0; i < upgrades.size(); i++) {
+                    Optional<SoldierUpgradeContainer> optionalSoldier = handleIncomingUpgradePacket(buf.entityId(), upgrades.get(i), context);
+                    if (optionalSoldier.isEmpty())
+                        return;
+                    SoldierUpgradeContainer container = optionalSoldier.get();
+                    container.soldier.upgrades.add(container.upgrade);
+                    container.upgrade.onAdd(container.soldier);
+                }
+            } else if (buf.action() == SyncUpgradesPayload.Action.REMOVE) {
+                List<ISoldierUpgrade> upgrades = buf.upgrades();
+                for (int i = 0; i < upgrades.size(); i++) {
+                    Optional<SoldierUpgradeContainer> optionalSoldier = handleIncomingUpgradePacket(buf.entityId(), upgrades.get(i), context);
+                    if (optionalSoldier.isEmpty())
+                        return;
+                    SoldierUpgradeContainer container = optionalSoldier.get();
+                    container.soldier.upgrades.remove(container.upgrade);
+                    container.upgrade.onRemove(container.soldier);
+                }
+            }
         });
     }
 
-    private static Optional<SoldierUpgradeContainer> handleIncomingUpgradePacket(PacketByteBuf buf, NetworkManager.PacketContext context) {
-        int soldierEntityId = buf.readInt();
-        Identifier upgradeIdentifier = buf.readIdentifier();
-        ISoldierUpgrade upgrade = UpgradeRegistry.SOLDIER_UPGRADE_REGISTER.get(upgradeIdentifier);
+    private static Optional<SoldierUpgradeContainer> handleIncomingUpgradePacket(int soldierEntityId, ISoldierUpgrade upgrade, NetworkManager.PacketContext context) {
         if (upgrade == null) {
-            ClayMod.LOGGER.warn("Unknown upgrade id {}", upgradeIdentifier);
+            ClayMod.LOGGER.warn("Unknown upgrade id");
             return Optional.empty();
         }
         Entity entity = context.getPlayer().getEntityWorld().getEntityById(soldierEntityId);
